feat#49: consolex -> console moved from internal tests to external
