+++ pkg/console/console_test.go
@@ -1,4 +1,4 @@
-package consolex
+package console
 
 import (
 	"bytes"
